+++ Rag_5040/BST.lua
@@ -2,20 +2,11 @@ local profile = {}
 
 local fastCastValue = 0.00 -- 0% from gear listed in Precast set
 
-local max_hp_in_idle_with_regen_gear_equipped = 1200
+local max_hp_in_idle_with_regen_gear_equipped = 9999 -- You could set this to 9999 if you do not wish to ever use regen gear
 
 -- Comment out the equipment within these sets if you do not have them or do not wish to use them
 local gaudy_harness = {
     -- Body = 'Gaudy Harness',
-}
-local muscle_belt = {
-    Waist = 'Muscle Belt +1',
-}
-local presidential_hairpin = {
-    -- Head = 'President. Hairpin',
-}
-local dream_ribbon = {
-    Head = 'Dream Ribbon',
 }
 
 local sets = {
@@ -233,20 +224,9 @@ profile.HandleCommand = function(args)
 end
 
 profile.HandleDefault = function()
-    gcmelee.DoDefault()
+    gcmelee.DoDefault(max_hp_in_idle_with_regen_gear_equipped)
 
     local player = gData.GetPlayer()
-    if (player.Status == 'Idle') then
-        if (player.HPP < 50) then
-            gFunc.EquipSet('muscle_belt')
-        end
-        if (player.HP < max_hp_in_idle_with_regen_gear_equipped) then
-            if (conquest:GetOutsideControl()) then
-                gFunc.EquipSet('presidential_hairpin')
-            end
-            gFunc.EquipSet('dream_ribbon')
-        end
-    end
     if (player.SubJob == 'NIN' and player.Status == 'Engaged') then
         gFunc.EquipSet('TP_NIN')
     end