+++ Rag_5040/NIN.lua
@@ -2,6 +2,8 @@ local profile = {}
 
 local fastCastValue = 0.00 -- 0% from gear listed in Precast set
 
+local max_hp_in_idle_with_regen_gear_equipped = 9999 -- You could set this to 9999 if you do not wish to ever use regen gear
+
 -- Comment out the equipment within these sets if you do not have them or do not wish to use them
 local fire_staff = {
     Main = 'Vulcan\'s Staff',
@@ -285,7 +287,7 @@ profile.HandleCommand = function(args)
 end
 
 profile.HandleDefault = function()
-    gcmelee.DoDefault()
+    gcmelee.DoDefault(max_hp_in_idle_with_regen_gear_equipped)
 
     local player = gData.GetPlayer()
     local environment = gData.GetEnvironment()