+++ Rag_5040/PLD.lua
@@ -2,6 +2,8 @@ local profile = {}
 
 local fastCastValue = 0.07 -- 7% from gear listed in Precast set
 
+local max_hp_in_idle_with_regen_gear_equipped = 9999 -- You could set this to 9999 if you do not wish to ever use regen gear
+
 -- Comment out the equipment within these sets if you do not have them or do not wish to use them
 local parade_gorget = {
     Neck = 'Parade Gorget',
@@ -573,7 +575,7 @@ profile.HandleCommand = function(args)
 end
 
 profile.HandleDefault = function()
-    gcmelee.DoDefault()
+    gcmelee.DoDefault(max_hp_in_idle_with_regen_gear_equipped)
 
     local player = gData.GetPlayer()
     local cover = gData.GetBuffCount('Cover')